# api/openapi/loom.yaml
openapi: 3.1.0
info:
  title: Loom API
  version: 0.1.0-alpha
  summary: Native HTTP API for Loom — unified media automation.
  description: |
    Loom's native API is served under `/api/v1/*`. Wire-compatible
    surfaces (`/api/v3/*` for Radarr / Sonarr, `/api/v1/*` for Prowlarr)
    are added in Phase 7 and are documented separately.

    This specification covers only the routes implemented today. The
    PR template requires every new route to be reflected here.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  contact:
    name: Loom maintainers
    url: https://github.com/ebenderooock/loom
servers:
  - url: http://localhost:8989
    description: Local dev server.
  - url: https://{host}
    description: Self-hosted instance behind a reverse proxy.
    variables:
      host:
        default: loom.example.com
...
paths:
...
  /api/v1/indexers/:
    get:
      tags: [indexers]
      summary: List configured indexers.
      operationId: listIndexers
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Wrapped list of indexer definitions, each with the latest health row.
          content:
            application/json:
              schema:
                type: object
                required: [indexers]
                properties:
                  indexers:
                    type: array
                    items:
                      $ref: "#/components/schemas/IndexerWithHealth"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [indexers]
      summary: Create an indexer.
      operationId: createIndexer
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexerCreate"
      responses:
        "201":
          description: Indexer created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indexer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/indexers/{id}:
    parameters:
      - $ref: "#/components/parameters/IndexerID"
    get:
      tags: [indexers]
      summary: Get an indexer.
      operationId: getIndexer
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Indexer definition with latest health row.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexerWithHealth"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [indexers]
      summary: Replace an indexer.
      operationId: replaceIndexer
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexerCreate"
      responses:
        "200":
          description: Updated indexer definition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indexer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [indexers]
      summary: Partially update an indexer.
      operationId: patchIndexer
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexerPatch"
      responses:
        "200":
          description: Updated indexer definition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indexer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [indexers]
      summary: Delete an indexer.
      operationId: deleteIndexer
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "204":
          description: Indexer deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/indexers/{id}/caps:
    parameters:
      - $ref: "#/components/parameters/IndexerID"
    get:
      tags: [indexers]
      summary: Capability descriptor for a single indexer.
      operationId: getIndexerCaps
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Indexer capabilities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caps"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/indexers/{id}/test:
    parameters:
      - $ref: "#/components/parameters/IndexerID"
    post:
      tags: [indexers]
      summary: Run an immediate health check; persists the result.
      operationId: testIndexer
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Outcome of the test plus measured latency. The same row is persisted into `indexer_health`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/indexers/search:
    post:
      tags: [indexers]
      summary: Fan-out search across enabled indexers.
      operationId: searchIndexers
      security:
        - apiKey: []
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Aggregated results plus per-source error map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregatedResults"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  schemas:
...
    Indexer:
      type: object
      required: [id, kind, name, enabled, priority, categories, tags]
      properties:
        id:
          type: string
          description: Slug identifier; generated from kind+name when omitted on create.
        kind:
          type: string
          description: |
            Factory key. Built-in kinds: `builtin/null`. Real source
            kinds: `newznab`, `torznab`, `cardigann`.
          enum:
            - builtin/null
            - newznab
            - torznab
            - cardigann
        name:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
          minimum: 0
          maximum: 100
        config:
          description: |
            Free-form, kind-specific configuration. The shape is
            documented per kind; for `newznab` and `torznab` it
            matches `NewznabConfig`, for `cardigann` it matches
            `CardigannConfig`.
          oneOf:
            - $ref: "#/components/schemas/NewznabConfig"
            - $ref: "#/components/schemas/CardigannConfig"
            - type: object
              additionalProperties: true
        categories:
          type: array
          description: Newznab category IDs.
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
        proxy_id:
          type: string
          description: Optional reference to a `proxies.id`. When set, all outbound HTTP traffic from this indexer is routed through the matching proxy row.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IndexerWithHealth:
      allOf:
        - $ref: "#/components/schemas/Indexer"
        - type: object
          properties:
            health:
              $ref: "#/components/schemas/IndexerHealth"
            rate_limit:
              $ref: "#/components/schemas/IndexerRateLimit"
    IndexerRateLimit:
      type: object
      description: |
        Per-indexer rate-limit + retry configuration. The first three
        fields reflect the overrides stored on the row (any may be
        absent); the `effective_*` fields are what the runtime is
        actually applying after defaults are merged in.
      required: [effective_per_minute, effective_burst, effective_max_retries]
      properties:
        per_minute:
          type: integer
          minimum: 1
          description: Override for the token-bucket fill rate (req/min). Absent when the row stores NULL.
        burst:
          type: integer
          minimum: 1
          description: Override for the token-bucket capacity. Absent when the row stores NULL.
        max_retries:
          type: integer
          minimum: 0
          description: Override for the retry budget. Absent when the row stores NULL.
        effective_per_minute:
          type: integer
          description: Per-minute fill rate the runtime is using.
        effective_burst:
          type: integer
          description: Burst capacity the runtime is using.
        effective_max_retries:
          type: integer
          description: Retry budget the runtime is using.
    IndexerCreate:
      type: object
      required: [kind, name]
      properties:
        id:
          type: string
        kind:
          type: string
        name:
          type: string
        enabled:
          type: boolean
          default: true
        priority:
          type: integer
          default: 25
        config:
          type: object
          additionalProperties: true
        categories:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
        proxy_id:
          type: string
          description: Optional `proxies.id` to attach on creation.
        rate_limit_per_min:
          type: integer
          minimum: 1
          description: Optional override for the token-bucket fill rate. Omit to use the package default (60 req/min).
        rate_limit_burst:
          type: integer
          minimum: 1
          description: Optional override for the token-bucket capacity. Omit to use the package default (5).
        retry_max_attempts:
          type: integer
          minimum: 0
          description: Optional override for the retry budget. Pass `0` to disable retries entirely; omit to use the package default (3).
    IndexerPatch:
      type: object
      description: Any subset of fields. Omitted fields are left untouched.
      properties:
        name:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        config:
          type: object
          additionalProperties: true
        categories:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
        proxy_id:
          type: string
          nullable: true
          description: Set to a proxy ID to pin outbound traffic. Send `null` (or empty string) to clear the existing pin.
        rate_limit_per_min:
          type: integer
          minimum: 1
          description: Set the token-bucket fill rate override. Omit to leave unchanged.
        rate_limit_burst:
          type: integer
          minimum: 1
          description: Set the token-bucket capacity override. Omit to leave unchanged.
        retry_max_attempts:
          type: integer
          minimum: 0
          description: Set the retry budget override. Pass `0` to disable retries; omit to leave unchanged.
    IndexerHealth:
      type: object
      required: [indexer_id, status, last_checked_at]
      properties:
        indexer_id:
          type: string
        status:
          type: string
          enum: [ok, degraded, failed, unknown]
        last_checked_at:
          type: string
          format: date-time
        last_success_at:
          type: string
          format: date-time
        latency_ms:
          type: integer
          description: Round-trip time of the last `Test()` call, in milliseconds.
        last_error:
          type: string
          description: Last error message, if any.
...
    TestResult:
      type: object
      required: [ok, latency_ms]
      properties:
        ok:
          type: boolean
        latency_ms:
          type: integer
        error:
          type: string
...
    Caps:
      type: object
      required: [search_types, categories, supported_ids]
      properties:
        search_types:
          type: array
          description: Newznab-style search modes (e.g. `search`, `tv-search`, `movie-search`).
          items:
            type: string
        categories:
          type: array
          items:
            type: integer
        supported_ids:
          type: array
          description: External ID schemes accepted in searches (e.g. `imdb`, `tvdb`, `tmdb`).
          items:
            type: string
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        categories:
          type: array
          items:
            type: integer
        indexer_ids:
          type: array
          description: Restrict the fan-out to this subset of indexer IDs.
          items:
            type: string
        timeout_ms:
          type: integer
          description: Per-indexer ceiling override for this call only.
...
    AggregatedResults:
      type: object
      required: [results, errors]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Result"
        errors:
          type: object
          description: Per-indexer error messages, keyed by indexer ID. Empty object when all sources succeeded.
          additionalProperties:
            type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              examples: [bad_request, not_found, conflict, internal]
            message:
              type: string
...
  parameters:
...
    IndexerID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request body or parameters could not be processed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: A resource with the same identifier already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Long-lived API key. Issued via the UI / `loom apikey` CLI in
        Phase 1c. Stored at rest as a SHA-256 hash plus a 4-character
        non-secret prefix.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token (placeholder for OIDC access tokens). Auth routes
        consuming this scheme land in Phase 1c.
    cookieAuth:
      type: apiKey
      in: cookie
      name: loom_session
      description: |
        HTTP-only session cookie issued by the forms-login flow.
        Auth routes consuming this scheme land in Phase 1c.

# Endpoints implemented today are intentionally unauthenticated; once
# the Phase 1c auth PR lands, security requirements will be applied
# at the operation level.
security: []